Remove ManifestCommand
